refactor: progressbar 로직 분리
