+++ src/app/chat/filter/page.tsx
@@ -5,10 +5,12 @@ import { useRouter } from "next/navigation";
 
 import useFunnel from "@/hooks/useFunnel";
 
+import { Header } from "@/layout/Header";
+
 import LocationStep from "@/components/setup/LocationStep";
 import OptionSelectStep from "@/components/setup/OptionSelectStep";
 import BudgeStep from "@/components/setup/BudgeStep";
-import Header from "@/layout/Header";
+import ProgressBar from "@/components/setup/ProgressBar";
 
 const TRANSACTION_OPTIONS = ["월세", "전세", "매매"];
 const HOUSING_OPTIONS = ["원룸 / 투룸", "빌라", "오피스텔", "아파트"];
@@ -17,35 +19,16 @@ const Page = () => {
   const { Funnel, Step, currentStep, prevStep, nextStep } = useFunnel({ lastStep: "4" });
   const router = useRouter();
 
-  const getProgressWidthClass = (step: string) => {
-    switch (step) {
-      case "1":
-        return "w-1/4";
-      case "2":
-        return "w-2/4";
-      case "3":
-        return "w-3/4";
-      case "4":
-        return "w-full";
-      default:
-        return "w-0";
-    }
-  };
-
   return (
     <div className="min-h-screen">
-      <Header
-        title={"필터 설정하기"}
-        onPrevClick={prevStep}
-        onCloseClick={() => router.push("/")}
-      />
+      <Header bgColorClassName="bg-gray-100">
+        <Header.Prev onPrevClick={prevStep} />
+        <Header.Title>필터 설정하기</Header.Title>
+        <Header.Close onCloseClick={() => router.push("/")} />
+      </Header>
       <main className="relative min-h-screen pt-16">
-        {/** Progress Bar */}
-        <div className="h-[4px] w-full bg-gray-300">
-          <div
-            className={`h-[4px] rounded-br-[16px] rounded-tr-[16px] bg-blue-800-primary transition-all duration-300 ${getProgressWidthClass(currentStep)}`}
-          />
-        </div>
+        <ProgressBar currentStep={currentStep} />
+
         <div className="px-5 pt-10">
           <Funnel>
             <Step name="1">
